+++ RazorEngineCms/DAL/Repository/Repository.cs
@@ -8,22 +8,6 @@ using System.Threading.Tasks;
 
 namespace RazorEngineCms.DAL.Repository
 {
-    public interface IRepository<T>
-    {
-        Page Find(string section, string name, DateTime? updated = null);
-
-        T Find(int Id);
-
-        List<T> All();
-
-        Task Save(T obj, ConcurrentBag<string> errors);
-
-        Task<T> Copy(T page, ConcurrentBag<string> errors);
-
-        Task Delete(T obj, ConcurrentBag<string> errors);
-
-    }
-
     public class Repository<T> : IRepository<T> where T: class
     {
         internal ApplicationContext _db { get; set; }
