+++ RazorEngineCms/Views/Page/Edit.cshtml
@@ -21,7 +21,7 @@
         <form id="newPage" action="">
             <div class="form-group">
                 <label for="pageId">Page Id:</label>
-                <span id="pageId">@Model.Id</span> <br/>
+                <span id="pageId">@Model.Id</span> <br />
                 <label for="pageUpdated">Updated:</label>
                 <span id="pageUpdated">@Model.Updated.ToString()</span>
             </div>
@@ -55,20 +55,14 @@
                     Pages can optionally have a page model that uses Url Parameters, User Info, or DataBase Info.
                 </p>
                 <label for="hasModel-true" class="weight-normal">
-                    <input type="radio" name="hasModel" id="hasModel-true" value="true" class="formControl" 
-                           @if(!string.IsNullOrEmpty(Model.Model))
-                           {
-                               @:checked
-                           }
-                            /> Yes
+                    <input type="radio" name="hasModel" id="hasModel-true" value="true" class="formControl"
+                           @if (!string.IsNullOrEmpty(Model.Model)) { @: checked
+                                                      } /> Yes
                 </label>
                 <label for="hasModel-false" class="weight-normal">
-                    <input type="radio" name="hasModel" id="hasModel-false" value="false" class="formControl" 
-                           @if(string.IsNullOrEmpty(Model.Model))
-                           {
-                               @:checked
-                           } 
-                           /> No
+                    <input type="radio" name="hasModel" id="hasModel-false" value="false" class="formControl"
+                           @if (string.IsNullOrEmpty(Model.Model)) { @: checked
+                                                      } /> No
                 </label>
             </div>
             <div class="well" id="pageModelSection" style="display:
@@ -88,17 +82,17 @@
                     <label for="hasParams-true" class="weight-normal">
                         <input type="radio" name="hasParams" id="hasParams-true" value="true" class="formControl"
                                @if (Model.HasParams) { @: checked="checked"
-                                                                                         } /> True
+                                                                                                                        } /> True
                     </label>
                     <label for="hasParams-false" class="weight-normal">
                         <input type="radio" name="hasParams" id="hasParams-false" value="false" class="formControl"
                                @if (Model.HasParams != true) { @: checked="checked"
-                                                                                         } /> False
+                                                                                                                        } /> False
                     </label>
                 </div>
                 <div class="form-group">
                     <label for="model">Page Model</label> (<a href="#" data-toggle="modal" data-target="#modelEditorInfoModal">More Info</a>)
-                    <p>Create an object variable named <code>Model</code> to pass .NET code to view.</p>
+                    <p>Create an object variable named <code>Model</code> and access it in view as <code>@@Model.PageModel</code>.</p>
                     <textarea class="form-control" id="model" name="model"></textarea>
                 </div>
             </div>
@@ -109,15 +103,15 @@
                 </p>
                 <label for="hasInclude-true" class="weight-normal">
                     <input type="radio" name="hasInclude" id="hasInclude-true" value="true" class="formControl"
-                            @if (Model.HasInclude == true) { @: checked="checked"
-                                                        } /> Yes
+                           @if (Model.HasInclude == true) { @: checked="checked"
+                                                                                   } /> Yes
                 </label>
                 <label for="hasInclude-false" class="weight-normal">
                     <input type="radio" name="hasInclude" id="hasInclude-false" value="false" class="formControl"
-                            @if (Model.HasInclude != true) { @: checked="checked"
-                                                        } /> No
+                           @if (Model.HasInclude != true) { @: checked="checked"
+                                                                                   } /> No
                 </label>
-            </div> 
+            </div>
             <div class="form-group">
                 <label for="template">Page Template </label>
                 <p>Access Page Model with Razor Syntax (e.g. <code>@@Model.PropertyName</code>)</p>
@@ -127,7 +121,7 @@
                 <input type="submit" class="btn btn-default" id="submitNewPage" />
                 <button type="button" class="btn btn-danger" id="deletePage">Delete Page</button>
             </div>
-</form>
+        </form>
     </div>
 </div>
 
@@ -163,7 +157,7 @@
             else
             {
                 @:''
-            });
+                    });
         pageTemplateEditor.setSize("100%", "80%");
         $(function () {
             if (!(page instanceof Page)) {
