+++ RazorEngineCms/Views/Page/New.cshtml
@@ -63,7 +63,7 @@
                     </div>
                     <div class="form-group">
                         <label for="model">Page Model</label> (<a href="#" data-toggle="modal" data-target="#modelEditorInfoModal">More Info</a>)
-                        <p>Create an object variable named <code>Model</code> and access it in view as <code>@@Model.PageMoel</code>.</p>
+                        <p>Create an object variable named <code>Model</code> and access it in view as <code>@@Model.PageModel</code>.</p>
                         <textarea class="form-control" id="model" name="model"></textarea>
                     </div>
                 </div>
